requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
PyMySQL==1.1.1
Werkzeug==3.1.3
python-dotenv==1.0.1